% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsplot.r
\name{rtsplot.candle.col}
\alias{rtsplot.candle.col}
\alias{rtsplot.volume.col}
\title{Bar Colors for Candle and Volume plots}
\usage{
rtsplot.candle.col(y)

rtsplot.volume.col(y)
}
\arguments{
\item{y}{\code{\link{xts}} object}
}
\value{
colors
}
\description{
Bar Colors for Candle and Volume plots
}
