% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsplot.r
\name{rtsplot.grid}
\alias{rtsplot.grid}
\title{Add grid to time series plot}
\usage{
rtsplot.grid(grid, xaxis.ticks, col = rtsplot.theme()$grid.color)
}
\arguments{
\item{grid}{which grid lines to draw, \strong{defaults to 'xy'}}

\item{xaxis.ticks}{location of x axis ticks}

\item{col}{grid color, \strong{defaults to rtsplot.theme()$grid.color}}
}
\value{
nothing
}
\description{
Add grid to time series plot
}
