% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsplot.r
\name{rtsplot.hl}
\alias{rtsplot.hl}
\title{Create HL Plot}
\usage{
rtsplot.hl(y, col = rtsplot.volume.col(y), border = rtsplot.theme()$col.border)
}
\arguments{
\item{y}{\code{\link{xts}} object}

\item{col}{color for bars, \strong{defaults to rtsplot.volume.col}}

\item{border}{border color, \strong{defaults to rtsplot.theme()$col.border}}
}
\value{
nothing
}
\description{
Create HL Plot
}
\examples{
y = rtsplot.fake.stock.data(50, ohlc=TRUE)
symbol = 'SPY'
	
 # plot
 layout(1)
 rtsplot(y, type = 'n')
 rtsplot.hl(y)
rtsplot.legend(symbol, 'black', y)

}
