% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsplot.r
\name{rtsplot.layout}
\alias{rtsplot.layout}
\title{Create layout}
\usage{
rtsplot.layout(ilayout, delim = ",")
}
\arguments{
\item{ilayout}{matrix stored as a string}

\item{delim}{delimiter, \strong{defaults to ','}}
}
\value{
nothing
}
\description{
Create layout
}
