% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsplot.r
\name{rtsplot.legend}
\alias{rtsplot.legend}
\title{Plot legend - shortcut to the \code{\link{legend}} function}
\usage{
rtsplot.legend(
  labels,
  fill = NULL,
  lastobs = NULL,
  x = "topleft",
  merge = FALSE,
  bty = "n",
  border = NA,
  yformat = rtsplot.format,
  cex = 1,
  ...
)
}
\arguments{
\item{labels}{legend labels}

\item{fill}{fill colors, \strong{defaults to NULL}}

\item{lastobs}{list of last observations, \strong{defaults to NULL}}

\item{x}{location of legend, \strong{defaults to 'topleft'}}

\item{merge}{merge, \strong{defaults to FALSE}, see \code{\link{legend}} function for more info}

\item{bty}{box, \strong{defaults to 'n'}, see \code{\link{legend}} function for more info}

\item{border}{border color, \strong{defaults to NA - no color}}

\item{yformat}{format Y values function, \strong{defaults to \code{\link{rtsplot.format}}}}

\item{cex}{font size, \strong{defaults to 1}}

\item{...}{other parameters to legend, see \code{\link{legend}} function for more info}
}
\value{
nothing
}
\description{
Plot legend - shortcut to the \code{\link{legend}} function
}
\examples{
y = rtsplot.fake.stock.data(1000)
symbol = 'SPY'
	
 # plot
 layout(1)
 rtsplot(y, type = 'l', col='black')
rtsplot.legend(symbol, 'black', y)

}
