% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsplot.r
\name{rtsplot.lines}
\alias{rtsplot.lines}
\title{Add lines to time series plot}
\usage{
rtsplot.lines(y, type = "l", col = graphics::par("col"), ...)
}
\arguments{
\item{y}{\code{\link{xts}} object}

\item{type}{line type, \strong{defaults to 'l'}, for more info see \code{\link{lines}}}

\item{col}{color, \strong{defaults to par('col')}}

\item{...}{additional parameters to the \code{\link{lines}}}
}
\value{
nothing
}
\description{
Add lines to time series plot
}
