% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsplot.r
\name{rtsplot.matplot}
\alias{rtsplot.matplot}
\title{\code{\link[graphics]{matplot}} version for \code{\link[xts]{xts}} object}
\usage{
rtsplot.matplot(
  y,
  dates = NULL,
  ylim = NULL,
  type = "l",
  cols = rtsplot.colors(ncol(y)),
  ...
)
}
\arguments{
\item{y}{\code{\link{xts}} object}

\item{dates}{subset of dates\strong{defaults to NULL}}

\item{ylim}{range on Y values, \strong{defaults to NULL}}

\item{type}{plot type, \strong{defaults to 'l'}, see \code{\link[graphics]{plot}} for details}

\item{cols}{colors}

\item{...}{additional parameters to the \code{\link[graphics]{matplot}}}
}
\value{
nothing
}
\description{
\code{\link[graphics]{matplot}} version for \code{\link[xts]{xts}} object
}
