% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsplot.r
\name{rtsplot.scale.volume}
\alias{rtsplot.scale.volume}
\title{Scale volume}
\usage{
rtsplot.scale.volume(y)
}
\arguments{
\item{y}{\code{\link{xts}} object}
}
\value{
adjusted y object
}
\description{
Scale volume
}
