% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsplot.r
\name{rtsplot.stacked}
\alias{rtsplot.stacked}
\title{Create Stacked plot}
\usage{
rtsplot.stacked(
  x,
  y,
  xlab = "",
  cols = rtsplot.colors(ncol(y)),
  type = c("l", "s"),
  flip.legend = FALSE,
  ...
)
}
\arguments{
\item{x}{dates object}

\item{y}{matrix with weights}

\item{xlab}{X label, \strong{defaults to ''}, for more info see \code{\link{plot}}}

\item{cols}{colors, \strong{defaults to colors \code{\link{rtsplot.theme}}}}

\item{type}{plot type: lines, step stairs c('l','s')}

\item{flip.legend}{flag to reverse legend order, \strong{defaults to FALSE}}

\item{...}{additional parameters to the \code{\link{plot}}}
}
\value{
nothing
}
\description{
Create Stacked plot
}
