% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsplot.r
\name{rtsplot.text}
\alias{rtsplot.text}
\title{Add text to time series plot}
\usage{
rtsplot.text(y, ...)
}
\arguments{
\item{y}{\code{\link{xts}} object}

\item{...}{additional parameters to the \code{\link{lines}}}
}
\value{
nothing
}
\description{
Add text to time series plot
}
\examples{
y = rtsplot.fake.stock.data(1000)
symbol = 'SPY'
	
 # plot
 layout(1)
 rtsplot(y, type = 'l', col='black')
 rtsplot.text(y[100], 'Text', col='red')
rtsplot.legend(symbol, 'black', y)

}
