% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsplot.r
\name{rtsplot.volume}
\alias{rtsplot.volume}
\title{Plot volume}
\usage{
rtsplot.volume(
  y,
  col = rtsplot.volume.col(y),
  border = rtsplot.theme()$col.border
)
}
\arguments{
\item{y}{\code{\link{xts}} object}

\item{col}{color for volume bars}

\item{border}{color for volume bars border}
}
\value{
nothing
}
\description{
Plot volume
}
