% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsplot.r
\name{rtsplot.x.highlight}
\alias{rtsplot.x.highlight}
\title{Highlight vertical segments}
\usage{
rtsplot.x.highlight(y, highlight, col = rtsplot.theme()$col.x.highlight)
}
\arguments{
\item{y}{\code{\link{xts}} object}

\item{highlight}{segments to highlight along X axis}

\item{col}{highlight color, \strong{defaults to rtsplot.control$col.x.highlight}}
}
\value{
nothing
}
\description{
Highlight vertical segments
}
