% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsplot.r
\name{register.theme}
\alias{register.theme}
\alias{rtsplot.theme}
\alias{rtsplot.theme.set}
\alias{rtsplot.colors}
\title{Theme}
\usage{
register.theme(
  grid.color = "gray90",
  colors = "Set1",
  col.border = "black",
  col.up = "green",
  col.dn = "red",
  col.x.highlight = "orange",
  col.y.highlight = "orange",
  cex = 1,
  legend.bg.col = grDevices::adjustcolor("white", 200/255)
)

rtsplot.theme()

rtsplot.theme.set(...)

rtsplot.colors(n)
}
\arguments{
\item{grid.color}{color for grid lines, \strong{defaults to 'gray90'}}

\item{colors}{RColorBrewer set to generate colors, \strong{defaults to "Set1" in RColorBrewer}}

\item{col.border}{border color for drawing candles, \strong{defaults to 'black'}}

\item{col.up}{up color for drawing candles, \strong{defaults to 'green'}}

\item{col.dn}{down color for drawing candles, \strong{defaults to 'red'}}

\item{col.x.highlight}{color for highlighting along x axis, \strong{defaults to 'orange'}}

\item{col.y.highlight}{color for highlighting along y axis, \strong{defaults to 'orange'}}

\item{cex}{font size, \strong{defaults to 1}}

\item{legend.bg.col}{background legend color, \strong{defaults to grDevices::adjustcolor('white', 200/255)}}

\item{...}{additional settings}

\item{n}{number of colors to generate}
}
\value{
None
}
\description{
Setup theme
}
